DDP-5720 Move activity statuses to the SDK
