+++ ddp-workspace/projects/ddp-rarex/src/app/components/dashboard/dashboard.component.ts
@@ -13,7 +13,7 @@ import {
 } from 'ddp-sdk';
 
 import { ActivityCodes } from '../../constants/activity-codes';
-import { ActivityStatusCodes } from '../../constants/activity-status-codes';
+import { ActivityStatusCodes } from '../../../../../ddp-sdk/src/lib/models/activity/activityStatusCodes';
 import { CurrentActivityService } from '../../services/current-activity.service';
 import { RoutePaths } from '../../router-resources';
